Move some functions to utils
