+++ src/App.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import { createStore } from 'redux'
 import { connect } from 'react-redux';
-
 import './App.css';
 import AppReducer from './AppReducer';
 import * as actions from './AppActions';
@@ -13,6 +12,7 @@ import KickDialog from './Components/KickDialogComponent/KickDialog';
 import LoadingScreen from './Components/LoadingScreenComponent/LoadingScreen';
 import CreditsMenu from './Components/CreditsMenuComponent/CreditsMenu';
 import AdminMenu from './Components/AdminMenuComponent/AdminMenu'
+import { rgbToHex, hexToRgb } from './utils';
 
 export const store = createStore(AppReducer);
 
@@ -304,19 +304,6 @@ export const getColor = (id) => {
     }
   }
   return "#000000";
-}
-
-export function rgbToHex(r, g, b) {
-  return "#" + ((1 << 24) + (r << 16) + (g << 8) + b).toString(16).slice(1);
-}
-
-export function hexToRgb(hex) {
-  var result = /^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})$/i.exec(hex);
-  return result ? {
-    r: parseInt(result[1], 16),
-    g: parseInt(result[2], 16),
-    b: parseInt(result[3], 16)
-  } : null;
 }
 
 export default App
